refactor: move debug panel widget and helpers
